requirements.txt
falcon
gunicorn
requests
dnspython

pytest